requirements.txt
flask
flask-login
flask-sqlalchemy
python-dotenv
pandas
plotly
openai
werkzeug
gunicorn